@@ -1,35 +1,47 @@
 {
   "name": "harcmap",
-  "version": "1.3.0",
+  "version": "2.0.0-dev",
   "main": "src/index.js",
   "scripts": {
     "build:dev": "npx webpack --config webpack.dev.js",
     "build:dev:watch": "npx webpack --config webpack.dev.js --watch",
+    "build:mobile": "npx webpack --config webpack.mobile.js",
+    "build:mobile:watch": "npx webpack --config webpack.mobile.js --watch",
     "build:prod": "npx webpack --config webpack.prod.js",
     "build:prod:watch": "npx webpack --config webpack.prod.js --watch",
     "import:sass:watch": "sass-watcher --config sass-watcher.json"
   },
   "dependencies": {
+    "@capacitor/android": "^3.0.1",
+    "@capacitor/app": "^1.0.1",
+    "@capacitor/cli": "^3.0.1",
+    "@capacitor/core": "^3.0.1",
+    "@capacitor/geolocation": "^1.0.1",
+    "@capacitor/ios": "^3.3.1",
+    "@capacitor/local-notifications": "^1.0.1",
     "@dbetka/sass-watcher": "^0.1.5",
     "@dbetka/utils": "^0.1.3",
-    "@dbetka/vue-material-icons": "^0.2.7",
+    "@dbetka/vue-material-icons": "^0.3.1",
     "@dbetka/wdk": "^0.1.5",
+    "@egjs/flicking-plugins": "^4.2.2",
+    "@egjs/vue3-flicking": "^4.4.1",
+    "@vuelidate/core": "^2.0.0-alpha.32",
+    "@vuelidate/validators": "^2.0.0-alpha.25",
     "dayjs": "^1.10.7",
     "js-cookie": "^2.2.1",
     "js-search": "^2.0.0",
+    "ol": "^6.9.0",
     "lazy-line-painter": "^1.9.6",
-    "ol": "^6.2.1",
-    "vee-validate": "^3.4.11",
-    "vue": "^2.6.14",
-    "vue-concise-slider": "^4.2.1",
-    "vue-ellipse-progress": "^1.3.0",
-    "vue-i18n": "^8.25.0",
-    "vue-router": "^3.5.2",
-    "vue2-touch-events": "^3.2.2",
-    "vuex": "^3.6.2"
+    "vue": "^3.2.21",
+    "vue-ellipse-progress": "^2.0.0",
+    "vue-i18n": "^9.2.0-beta.17",
+    "vue-router": "^4.0.12",
+    "vue3-touch-events": "^4.1.0",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.14.8",
+    "@vue/compiler-sfc": "^3.2.21",
     "babel-loader": "^8.2.2",
     "babel-plugin-transform-builtin-extend": "^1.1.2",
     "copy-webpack-plugin": "^9.0.1",
@@ -40,10 +52,9 @@
     "sass-loader": "^12.1.0",
     "style-loader": "^3.2.1",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.4.3",
+    "typescript": "^4.5.2",
     "url-loader": "^4.1.1",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.14",
+    "vue-loader": "^16.8.3",
     "webpack": "^5.49.0",
     "webpack-cli": "^4.8.0",
     "webpack-merge": "^5.8.0"
