@@ -1,25 +1,23 @@
 <template>
-  <div id="app">
-    <o-header/>
-    <div class="f-relative f-flex-1">
-      <router-view :key="routerId"/>
-    </div>
-    <o-footer
-      v-touch:swipe.left="openMenuIfLogin"
-      v-touch:swipe.right="closeMenu"
-    />
-    <o-menu/>
-    <div
-      class="a-cover f-menu"
-      :class="isOpen ? 'f-show' : ''"
-      @click="closeMenu"
-      v-touch:swipe.right="closeMenu"
-    />
-    <o-popup/>
-    <m-snackbar/>
-    <o-guide/>
-    <o-loading/>
+  <o-header/>
+  <div class="f-relative f-flex-1">
+    <router-view :key="routerId"/>
   </div>
+  <o-footer
+    v-touch:swipe.left="openMenuIfLogin"
+    v-touch:swipe.right="closeMenu"
+  />
+  <o-menu/>
+  <div
+    class="a-cover f-menu"
+    :class="isOpen ? 'f-show' : ''"
+    @click="closeMenu"
+    v-touch:swipe.right="closeMenu"
+  />
+  <o-popup/>
+  <m-snackbar/>
+  <o-guide/>
+  <o-loading/>
 </template>
 
 <script>
@@ -44,7 +42,6 @@ export default {
   },
   computed: {
     ...mapGetters([
-      'isLoading',
       'routerId',
     ]),
     ...mapGetters('menu', [
