@@ -20,7 +20,7 @@
       v-for="(route, key) in links"
       :key="key"
       :to="route.path"
-      @click.native="close()"
+      @click="close()"
       class="a-link f-menu"
       :class="{ 'f-selected': isActualPath(route) }"
     >
@@ -34,19 +34,19 @@
     <a-link-menu
       v-if="checkIsCommon()"
       @click="openGuide()"
-      :icon="ICONS.help"
+      :icon="$icons.names.help"
       :text="$t('features.guide.howAppWorks')"
     />
 
     <a-link-menu
       @click="toggleTheme()"
-      :icon="ICONS.brightness_4"
+      :icon="$icons.names.brightness_4"
       :text="themeName === THEMES.light ? $t('general.darkTheme') : $t('general.lightTheme')"
     />
 
     <a-link-menu
       @click="signOut()"
-      :icon="ICONS.logout"
+      :icon="$icons.names.logout"
       :text="$t('general.logout')"
     />
 
@@ -90,6 +90,7 @@ export default {
         ROUTES.map,
         isCommon ? ROUTES.about : undefined,
         ROUTES.eventsList,
+        ROUTES.teamView,
       ];
       return links.filter(route => uCheck.isUndefined(route) === false);
     },
